pyproject.toml
[project]
name = "giftginnie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-environ>=0.11.2",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "psycopg2>=2.9.10",
    "django-cors-headers>=4.6.0",
    "ruff>=0.8.4",
    "requests>=2.32.3",
    "cloudinary>=1.42.1",
]
